use gorm

+++ services/order_service.go
@@ -6,10 +6,10 @@ import (
 )
 
 type IOrderService interface {
-	GetOrderByID(int64) (*datamodels.Order, error)
-	DeleteOrderByID(int64) bool
+	GetOrderByID(uint) (*datamodels.Order, error)
+	DeleteOrderByID(uint) bool
 	UpdateOrder(*datamodels.Order) error
-	InsertOrder(*datamodels.Order) (int64, error)
+	InsertOrder(*datamodels.Order) (uint, error)
 	GetAllOrder() ([]*datamodels.Order, error)
 	GetAllOrderInfo() (map[int]map[string]string, error)
 }
@@ -22,11 +22,11 @@ func NewOrderService(repository repositories.IOrderRepository) IOrderService {
 	return &OrderService{OrderRepository: repository}
 }
 
-func (o *OrderService) GetOrderByID(orderId int64) (order *datamodels.Order, err error) {
+func (o *OrderService) GetOrderByID(orderId uint) (order *datamodels.Order, err error) {
 	return o.OrderRepository.SelectByKey(orderId)
 }
 
-func (o *OrderService) DeleteOrderByID(orderId int64) bool {
+func (o *OrderService) DeleteOrderByID(orderId uint) bool {
 	return o.OrderRepository.Delete(orderId)
 }
 
@@ -34,7 +34,7 @@ func (o *OrderService) UpdateOrder(order *datamodels.Order) error {
 	return o.OrderRepository.Update(order)
 }
 
-func (o *OrderService) InsertOrder(order *datamodels.Order) (int64, error) {
+func (o *OrderService) InsertOrder(order *datamodels.Order) (uint, error) {
 	return o.OrderRepository.Insert(order)
 }
